+++ robots/pkg/kubevirt/copy/jobs.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package jobs
+package copy
 
 import (
 	"context"
