+++ robots/pkg/kubevirt/require/BUILD.bazel
@@ -1,12 +1,30 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["require.go"],
+    srcs = [
+        "presubmits.go",
+        "require.go",
+    ],
     importpath = "kubevirt.io/project-infra/robots/pkg/kubevirt/require",
     visibility = ["//visibility:public"],
     deps = [
-        "//robots/pkg/kubevirt/require/presubmits:go_default_library",
+        "//robots/pkg/kubevirt/flags:go_default_library",
+        "//robots/pkg/kubevirt/github:go_default_library",
+        "//robots/pkg/kubevirt/log:go_default_library",
+        "//robots/pkg/querier:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
+        "@io_k8s_sigs_yaml//:go_default_library",
+        "@io_k8s_test_infra//prow/config:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["presubmits_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//robots/pkg/querier:go_default_library",
+        "@io_k8s_test_infra//prow/config:go_default_library",
     ],
 )
